requirements.txt
beautifulsoup4==4.12.3
flask==3.1.0
pandas==2.2.3
pip-chill==1.0.3
python-dotenv==1.0.1
requests==2.32.3
Flask-SQLAlchemy==3.1.1
